tpch-sim/tpch-sim/requirements.txt
faker
kafka-python